+++ src/mcp_runtime_server/test_runners/__init__.py
@@ -1,1 +1,2 @@
 """Test runners package initialization."""
+"""Test runners package initialization."""
